:ewfewf:

+++ client.c
@@ -1,48 +1,66 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   client.c                                           :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: zhaddoum <marvin@42.fr>                    +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2022/04/27 17:34:37 by zhaddoum          #+#    #+#             */
+/*   Updated: 2022/04/28 15:56:11 by zhaddoum         ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include<stdio.h>
 #include<unistd.h>
 #include<signal.h>
 #include<sys/wait.h>
-#include "LIBFT_42/libft.h"
+#include "libft/libft.h"
 
-	
+int	checkpid(char *pid, int arg)
+{
+	int	i;
+
+	i = -1;
+	if (arg != 3)
+	{
+		ft_putstr("wrong amount of arguments or wrong PID");
+		return (1);
+	}
+	while (pid[++i])
+	{
+		if (ft_isalpha(pid[i]) == 1)
+		{
+			ft_putstr("wrong PID!");
+			return (1);
+		}
+	}
+	return (0);
+}
+
 int	main(int argc, char **argv)
 {
-   int		i;
-   int      pid;
-   int		bit;
-   char 	*msg;
-	
-	pid = atoi(argv[1]);
-  	bit = 0;
+	int			i;
+	int			pid;
+	static int	bit;
+	char		*msg;
+
+	if (checkpid(argv[1], argc) == 1)
+		return (1);
 	i = 0;
+	pid = atoi(argv[1]);
 	msg = argv[2];
-//    if (argc != 2)
-//    	return (1);
-  	while(msg[i])
+	while (msg[i])
 	{
 		while (bit < 8)
 		{
 			if (msg[i] & 128 >> bit)
-			{
 				kill(pid, SIGUSR1);
-				usleep(100);
-			}
 			else
-			{
-				kill(pid,SIGUSR2);
-				usleep(100);
-			}
+				kill(pid, SIGUSR2);
 			bit++;
+			usleep(500);
 		}
 		bit = 0;
 		i++;
-	}
-	usleep(100);
-	while (bit < 8)
-	{
-		//printf("%d	",bit);
-		kill(pid,SIGUSR2);
-		bit++;
-		usleep(40);
 	}
 }
